# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vasooli API",
    "description": "AI-first revenue recovery — one policy layer, one audit trail, three engines.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/audit/batches/{batch_id}/summary": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Batch Summary",
        "description": "The headline metrics for one batch, recomputed from its entries.\n\nRecomputed rather than read from the batch's stored snapshot, so the number\nreported here and the trail behind it cannot disagree.",
        "operationId": "batch_summary_api_v1_audit_batches__batch_id__summary_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/mandate-recovery/runs": {
      "post": {
        "tags": [
          "mandate-recovery"
        ],
        "summary": "Trigger Run",
        "description": "Run the full loop over a mandate book and return the honest summary.",
        "operationId": "trigger_run_api_v1_mandate_recovery_runs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__api__v1__routes__mandate_recovery__RunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__engines__mandate_recovery__schemas__RunSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "mandate-recovery"
        ],
        "summary": "List Runs",
        "description": "Every mandate-recovery run, newest first. Each carries its seed.",
        "operationId": "list_runs_api_v1_mandate_recovery_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BatchRunRead"
                  },
                  "title": "Response List Runs Api V1 Mandate Recovery Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mandate-recovery/runs/{batch_id}": {
      "get": {
        "tags": [
          "mandate-recovery"
        ],
        "summary": "Get Run",
        "description": "One run's record, including the dataset batch id it consumed.",
        "operationId": "get_run_api_v1_mandate_recovery_runs__batch_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchRunRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mandate-recovery/runs/{batch_id}/summary": {
      "get": {
        "tags": [
          "mandate-recovery"
        ],
        "summary": "Get Run Summary",
        "description": "The run's own summary — the engine-specific figures the dashboard reads.\n\nServed from what the run stored rather than recomputed, because the splits in\nit (recovery by failure class, the AFA branch, wasted attempts avoided) need\nthe run's outcomes and the dataset behind them, and nothing serving an API\nrequest should be re-running an engine.\n\nThe money figures inside it were read off the audit trail at the end of the\nrun, and `/audit/batches/{batch_id}/summary` recomputes those independently\n— so the stored copy is checkable against the trail rather than merely\nconvenient. Quote that route for a headline; quote this one for the breakdown.",
        "operationId": "get_run_summary_api_v1_mandate_recovery_runs__batch_id__summary_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__engines__mandate_recovery__schemas__RunSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mandate-recovery/runs/{batch_id}/mandates": {
      "get": {
        "tags": [
          "mandate-recovery"
        ],
        "summary": "List Mandates",
        "description": "Every mandate in a run with its current recovery state and next step.",
        "operationId": "list_mandates_api_v1_mandate_recovery_runs__batch_id__mandates_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "route",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by derived failure route.",
              "title": "Route"
            },
            "description": "Filter by derived failure route."
          },
          {
            "name": "next_step",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by the decided next step.",
              "title": "Next Step"
            },
            "description": "Filter by the decided next step."
          },
          {
            "name": "afa_side",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`above` or `at_or_below` the Rs 15,000 AFA threshold.",
              "title": "Afa Side"
            },
            "description": "`above` or `at_or_below` the Rs 15,000 AFA threshold."
          },
          {
            "name": "compliance_blocked",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only mandates a compliance gate refused.",
              "title": "Compliance Blocked"
            },
            "description": "Only mandates a compliance gate refused."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MandateRecoveryStateRead"
                  },
                  "title": "Response List Mandates Api V1 Mandate Recovery Runs  Batch Id  Mandates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mandate-recovery/runs/{batch_id}/mandates/{mandate_id}": {
      "get": {
        "tags": [
          "mandate-recovery"
        ],
        "summary": "Get Mandate Timeline",
        "description": "One mandate's full timeline: every attempt, decision, rule and message.\n\nComplete on purpose. If a judge picks a mandate and asks \"why did it do that,\nand what stopped it going further?\", the answer is this one response — not a\nreconstruction from three endpoints.",
        "operationId": "get_mandate_timeline_api_v1_mandate_recovery_runs__batch_id__mandates__mandate_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "mandate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mandate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MandateTimeline"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mandate-recovery/runs/{batch_id}/communications": {
      "get": {
        "tags": [
          "mandate-recovery"
        ],
        "summary": "List Communications",
        "description": "Every message the run drafted. **None were sent** — see ADR 0009.",
        "operationId": "list_communications_api_v1_mandate_recovery_runs__batch_id__communications_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`marked_for_delivery`, `held` or `suppressed`. Never `sent`: nothing here dispatches.",
              "title": "Status"
            },
            "description": "`marked_for_delivery`, `held` or `suppressed`. Never `sent`: nothing here dispatches."
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MandateCommunicationRead"
                  },
                  "title": "Response List Communications Api V1 Mandate Recovery Runs  Batch Id  Communications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mandate-recovery/runs/{batch_id}/actions": {
      "get": {
        "tags": [
          "mandate-recovery"
        ],
        "summary": "List Actions",
        "description": "Every action the run took or refused, each naming its authorising rule.",
        "operationId": "list_actions_api_v1_mandate_recovery_runs__batch_id__actions_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEntryRead"
                  },
                  "title": "Response List Actions Api V1 Mandate Recovery Runs  Batch Id  Actions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchRunRead": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "engine": {
            "$ref": "#/components/schemas/Engine"
          },
          "seed": {
            "type": "integer",
            "title": "Seed"
          },
          "status": {
            "$ref": "#/components/schemas/BatchStatus"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "notes": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "engine",
          "seed",
          "status",
          "started_at"
        ],
        "title": "BatchRunRead"
      },
...
      "BatchSummary": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "entries": {
            "type": "integer",
            "title": "Entries"
          },
          "amount_at_risk_paise": {
            "type": "integer",
            "title": "Amount At Risk Paise"
          },
          "amount_recovered_paise": {
            "type": "integer",
            "title": "Amount Recovered Paise"
          },
          "recovery_rate": {
            "type": "number",
            "title": "Recovery Rate"
          },
          "action_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Action Counts"
          },
          "outcome_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Outcome Counts"
          },
          "engine_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Engine Counts"
          },
          "by_source": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SourceBreakdown"
            },
            "type": "object",
            "title": "By Source"
          },
          "policy_violations": {
            "type": "integer",
            "title": "Policy Violations"
          },
          "blocked_count": {
            "type": "integer",
            "title": "Blocked Count"
          },
          "halted_count": {
            "type": "integer",
            "title": "Halted Count"
          },
          "escalated_count": {
            "type": "integer",
            "title": "Escalated Count"
          },
          "abstained_count": {
            "type": "integer",
            "title": "Abstained Count"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "entries",
          "amount_at_risk_paise",
          "amount_recovered_paise",
          "recovery_rate",
          "action_counts",
          "outcome_counts",
          "engine_counts",
          "by_source",
          "policy_violations",
          "blocked_count",
          "halted_count",
          "escalated_count",
          "abstained_count"
        ],
        "title": "BatchSummary",
        "description": "The headline number, computed honestly from the log.\n\nNever from a separately maintained counter — a counter can drift from the\ntrail it claims to summarise, and then the demo's headline figure and its\nevidence disagree."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MandateTimeline": {
        "properties": {
          "mandate_id": {
            "type": "string",
            "title": "Mandate Id"
          },
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "state": {
            "$ref": "#/components/schemas/MandateRecoveryStateRead"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/AuditEntryRead"
            },
            "type": "array",
            "title": "Entries"
          },
          "communications": {
            "items": {
              "$ref": "#/components/schemas/MandateCommunicationRead"
            },
            "type": "array",
            "title": "Communications"
          },
          "next_step_summary": {
            "type": "string",
            "title": "Next Step Summary"
          }
        },
        "type": "object",
        "required": [
          "mandate_id",
          "batch_id",
          "state",
          "entries",
          "communications",
          "next_step_summary"
        ],
        "title": "MandateTimeline",
        "description": "One mandate's complete history — the route the demo leans on."
      },
...
      "SourceBreakdown": {
        "properties": {
          "entries": {
            "type": "integer",
            "title": "Entries",
            "default": 0
          },
          "amount_at_risk_paise": {
            "type": "integer",
            "title": "Amount At Risk Paise",
            "default": 0
          },
          "amount_recovered_paise": {
            "type": "integer",
            "title": "Amount Recovered Paise",
            "default": 0
          },
          "recovery_rate": {
            "type": "number",
            "title": "Recovery Rate",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "SourceBreakdown",
        "description": "Per-source figures.\n\nMetrics are reported *per source*, never blended: a recovery rate where half\nthe decisions came from static templates is not false, but as one figure it\nimplies more than it delivers."
      },
...
      "app__api__v1__routes__mandate_recovery__RunRequest": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id",
            "description": "Id for this engine run. Not the dataset's batch id."
          },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "default": 42
          },
          "dataset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset",
            "description": "Path to a mandates .jsonl. Defaults to the committed sample."
          },
          "now": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Now"
          }
        },
        "type": "object",
        "required": [
          "batch_id"
        ],
        "title": "RunRequest",
        "description": "What to run, and against what.\n\n`now` is exposed because the mandate book is anchored to a fixed `as_of` that\nis not today, and because the outreach window makes the run clock genuinely\nload-bearing: a batch executed at 03:00 IST correctly holds every customer\nmessage, which is right and looks wrong. Left unset, the runner derives the\nclock from the latest debit actually attempted in the book."
      },
...
    }
  }
}